+++ flureadium_platform_interface/test/extensions/url_test.dart
@@ -7,11 +7,29 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   test('remove last component', () {
-    expect(Uri.parse('http://domain.com/two/'), Uri.parse('http://domain.com/two/paths').removeLastComponent());
-    expect(Uri.parse('http://domain.com/two/'), Uri.parse('http://domain.com/two/paths/').removeLastComponent());
-    expect(Uri.parse('http://domain.com/'), Uri.parse('http://domain.com/path').removeLastComponent());
-    expect(Uri.parse('http://domain.com/'), Uri.parse('http://domain.com/path/').removeLastComponent());
-    expect(Uri.parse('http://domain.com/'), Uri.parse('http://domain.com/').removeLastComponent());
-    expect(Uri.parse('http://domain.com'), Uri.parse('http://domain.com').removeLastComponent());
+    expect(
+      Uri.parse('http://domain.com/two/'),
+      Uri.parse('http://domain.com/two/paths').removeLastComponent(),
+    );
+    expect(
+      Uri.parse('http://domain.com/two/'),
+      Uri.parse('http://domain.com/two/paths/').removeLastComponent(),
+    );
+    expect(
+      Uri.parse('http://domain.com/'),
+      Uri.parse('http://domain.com/path').removeLastComponent(),
+    );
+    expect(
+      Uri.parse('http://domain.com/'),
+      Uri.parse('http://domain.com/path/').removeLastComponent(),
+    );
+    expect(
+      Uri.parse('http://domain.com/'),
+      Uri.parse('http://domain.com/').removeLastComponent(),
+    );
+    expect(
+      Uri.parse('http://domain.com'),
+      Uri.parse('http://domain.com').removeLastComponent(),
+    );
   });
 }